refactor(#2454): extract Slice 12 tests to sibling file (anti-monolith ceiling)
mcp_proxy.rs grew to 2937 LOC (ceiling 2500) from the Slice 12 RED tests.
Extract all #2454 tests (1343 lines) into mcp_proxy_2454_tests.rs as a
#[cfg(test)] sibling module, keeping mcp_proxy.rs at 1597 LOC.

- invoke_runtime_mcp_tool made pub(super) for sibling visibility
- No test behavior change; all 54 #2454 tests pass


Agend-Agent: archfix-opus-3
Agend-Task: t-20260719050213255014-80319-1
Agend-Branch: fix/2454-send-runtime
Agend-Issued-At: 2026-07-19T08:39:35.183770+00:00
